Greatly improving readline and parseline performance
- To avoid allocate and free memory on every read and parse
operation, a buffer were included in both functions. Memory
will be allocated only if the buffer is NULL. This way you
can pass the buffer to the function and it will have memory
allocated only in the first read/parse, and after that the
buffer will be overwritten. You need to free the memory only
once, after fully parsing your file.